butter_app/requirements.txt
Django==3.2.11
django_filter==21.1
djangorestframework==3.13.1
gunicorn==20.0.4
celery==4.3.1
psycopg2-binary==2.8.4
redis==3.3.11

